(長田)component化
