+++ nagata_souji_train/data/duty.json
@@ -20,8 +20,8 @@
     "長田",
     "千々松"
   ],
-  "currentIndex": 0,
-  "lastUpdated": "2026-01-26T11:56:45.615Z",
+  "currentIndex": 17,
+  "lastUpdated": "2026-01-29T03:25:55.159Z",
   "history": [
     {
       "member": "中村",
@@ -5294,6 +5294,74 @@
     {
       "member": "千々松",
       "date": "2026-01-26T11:56:45.615Z"
+    },
+    {
+      "member": "石飛",
+      "date": "2026-01-29T03:00:22.673Z"
+    },
+    {
+      "member": "中土",
+      "date": "2026-01-29T03:00:23.787Z"
+    },
+    {
+      "member": "野村",
+      "date": "2026-01-29T03:00:27.529Z"
+    },
+    {
+      "member": "西川",
+      "date": "2026-01-29T03:00:29.066Z"
+    },
+    {
+      "member": "道田",
+      "date": "2026-01-29T03:00:30.425Z"
+    },
+    {
+      "member": "CHOI",
+      "date": "2026-01-29T03:00:31.312Z"
+    },
+    {
+      "member": "岡田",
+      "date": "2026-01-29T03:00:32.787Z"
+    },
+    {
+      "member": "江藤",
+      "date": "2026-01-29T03:00:33.986Z"
+    },
+    {
+      "member": "KIM",
+      "date": "2026-01-29T03:00:35.328Z"
+    },
+    {
+      "member": "ALIF",
+      "date": "2026-01-29T03:00:36.572Z"
+    },
+    {
+      "member": "中村",
+      "date": "2026-01-29T03:00:47.051Z"
+    },
+    {
+      "member": "市山",
+      "date": "2026-01-29T03:25:45.681Z"
+    },
+    {
+      "member": "中川",
+      "date": "2026-01-29T03:25:47.561Z"
+    },
+    {
+      "member": "小田",
+      "date": "2026-01-29T03:25:49.713Z"
+    },
+    {
+      "member": "向原",
+      "date": "2026-01-29T03:25:51.602Z"
+    },
+    {
+      "member": "一瀬",
+      "date": "2026-01-29T03:25:53.381Z"
+    },
+    {
+      "member": "稲澤",
+      "date": "2026-01-29T03:25:55.159Z"
     }
   ],
   "profiles": {
